Introduce sales view
A new sales view has been added that can let users filter by tags
customer, sector and some other fields. This is useful because when
responding to bids we can see what projects fit certain themes.

To achieve this a bunch of refactoring was done to make it easier to
add new code. Files were moved to more appropriate locations, and controllers
were split into their functional areas.
