+++ app/views/index.html
@@ -1,7 +1,7 @@
 {% extends "layout.html" %} {% block content %}
 <main id="content" role="main">
     <div class="grid-row">
-        <section class="column-two-thirds" />
+        <section class="column-two-thirds">
         <p>
             Our services, ordered by {{ view }}
             <!-- {% if view != 'agency' %}(<a href="/agency">order by agency</a>){% endif %} -->
